requirements.txt
Django==6.0.2
psycopg2==2.9.11
djangorestframework==3.16.1
django-filter==25.2
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.5.1